Add plural format implementation.
This provides the ability to write a classifier function which
signifies which slot on the format object should be used. For
English, this is "one" or "other". Other languages and uses will
use additional slots.

The set of slots available (the category) is based on what is
available in ICU as well as what Unicode defines for its pluralization
support. The machinery to implement your own classifier (or even
to specify one) is not yet provided.

* plural-format.dylan
  (<plural-format>): Provides slots for each possible format and the
   classifier function.
  (<plural-category>): The valid categories that a value can fall
   within.
  (english-cardinal-classifier): Default classifier implementation.
  (format-message-part): Basic implementation. This isn't really
   safe yet.

* tests/message-format-test-suite.dylan
  (basic-plural-format-test): Basic test.
  (message-format-test-suite): Run basic-plural-format-test.